feat: consolidate v1 and v2 of filter pages
