+++ packages/paste-website/src/components/ArticleList.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {Stack} from '@twilio-paste/stack';
 import {useUID} from '@twilio-paste/uid-library';
 import {Article} from './Article';
-import type {ArticleNode} from '../../@types/Article';
+import type {ArticleNode} from '../../types/Article';
 
 export interface ArticleListProps {
   articles: [
