+++ packages/paste-website/src/components/tokens-list/hooks/__tests__/hooks.test.ts
@@ -1,0 +1,10 @@
+// import {useThemeSettings, useTokenValueFormatter, useVirtualizedTokensList} from '../../hooks';
+
+// // @TODO tests
+// describe('Tokens hooks', () => {
+//   describe('useThemeSettings', () => {});
+
+//   describe('useTokenValueFormatter', () => {});
+
+//   describe('useVirtualizedTokensList', () => {});
+// });
