+++ packages/paste-website/src/components/tokens-list/hooks/index.ts
@@ -1,3 +1,3 @@
 export {useThemeSettings} from './useThemeSettings';
 export {useTokenValueFormatter} from './useTokenValueFormatter';
-export {useCopyTokenName, CopyIconWithTooltip} from './useCopyTokenName';
+export {useVirtualizedTokensList} from './useVirtualizedTokensList';
