+++ packages/paste-website/src/components/tokens-list/utils/__tests__/utils.test.ts
@@ -1,0 +1,14 @@
+// import {getTokenValue, isMatch, normalizeSearchString, sentenceCase, sortTokens} from '../../utils';
+
+// // @TODO update tests.
+// describe.skip('Tokens Utils', () => {
+//   describe('getTokenValue', () => {});
+
+//   describe('isMatch', () => {});
+
+//   describe('normalizeSearchString', () => {});
+
+//   describe('sentenceCase', () => {});
+
+//   describe('sortTokens', () => {});
+// });
