+++ packages/paste-website/src/components/tokens-list/utils/getTokenValue.ts
