+++ packages/paste-website/src/components/tokens-list/utils/googleAnalytics.ts
