+++ packages/paste-website/src/components/tokens-list/utils/normalizeSearchString.ts
