+++ packages/paste-website/src/components/tokens-list/utils/sentenceCase.ts
