--- packages/paste-website/src/pages/tokens/list/v2/index.mdx
@@ -1,16 +1,0 @@
----
-title: Design tokens WIP
-description: List of visual attributes (background colors, spacing, etc.) that our design language is based on.
-slug: /tokens/list/v2
----
-
-import {TokensList} from '../../../../components/tokens-list-v2';
-import {SidebarCategoryRoutes} from '../../../../constants';
-
-<ComponentHeader
-  name="Design Tokens"
-  categoryRoute={SidebarCategoryRoutes.TOKENS}
-  description="All tokens that are available in the Paste token system."
-/>
-
-<TokensList />
