+++ packages/paste-website/src/utils/color-contrast/__tests__/color-contrast.test.ts
@@ -1,0 +1,12 @@
+import {isLightToken, computeContrastRating, getAccessibilityData, getContrastRating} from '../../color-contrast';
+
+// @TODO tests
+describe('color contrast utilities', () => {
+  describe('isLightToken', () => {});
+
+  describe('computeContrastRating', () => {});
+
+  describe('getAccessibilityData', () => {});
+
+  describe('getContrastRating', () => {});
+});
