+++ packages/paste-website/src/utils/token-decorator/__tests__/token-decorator.test.ts
@@ -1,0 +1,12 @@
+// import {checkIsInverse, getTokensFromTheme, backgroundColors, annotate} from '../../token-decorator';
+
+// @TODO tests
+// describe('token data decorator', () => {
+//   describe('checkIsInverse', () => {});
+
+//   describe('getTokensFromTheme', () => {});
+
+//   describe('backgroundColors', () => {});
+
+//   describe('annotate', () => {});
+// });
