+++ packages/paste-website/tsconfig.json
@@ -16,8 +16,8 @@
     "skipLibCheck": true /* Skip type checking all .d.ts files. */,
     "outDir": "dist/",
     "emitDeclarationOnly": false,
-    "typeRoots": ["../../@types", "../../node_modules/@types", "./@types"]
+    "typeRoots": ["../../@types", "../../node_modules/@types", "./types"]
   },
-  "include": ["../../package.json", "./src/**/*", "./@types/**/*", "./static/**/*"],
+  "include": ["../../package.json", "./src/**/*", "./types/**/*", "./static/**/*"],
   "exclude": ["node_modules", "public", ".cache"]
 }
